pyproject.toml
[project]
name = "podking"
version = "0.1.0"
description = "Personal YouTube/podcast summarizer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pgvector>=0.3.6",
    "authlib>=1.3",
    "httpx>=0.27",
    "itsdangerous>=2.2",
    "pydantic-settings>=2.6",
    "cryptography>=43",
    "structlog>=24.4",
    "python-multipart>=0.0.19",
    "psycopg2-binary>=2.9",
    "yt-dlp>=2024.1.1",
    "feedparser>=6.0",
    "anthropic>=0.40",
    "elevenlabs>=1.0",
    "voyageai>=0.3",
    "sse-starlette>=2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
# B008 flags Depends(...) in FastAPI default args, which is idiomatic.
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = ["pgvector.*", "feedparser", "voyageai", "voyageai.*"]
ignore_missing_imports = true

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["backend/podking"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
